+++ src/js/extensions/snp_attestation.cpp
@@ -103,8 +103,12 @@ namespace ccf::js::extensions
 
       try
       {
-        verified_attestation.emplace(pal::verify_snp_attestation_report_and_get(
-          quote_info, measurement, report_data));
+        verified_attestation.emplace(pal::snp::AttestationReport::verify(
+          quote_info.quote,
+          quote_info.endorsements,
+          measurement,
+          report_data,
+          quote_info.endorsed_tcb));
         if (uvm_endorsements.has_value())
         {
           parsed_uvm_endorsements =